backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
yfinance==0.2.52
pydantic==2.7.1
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.2.1
python-dotenv==1.0.1
APScheduler==3.10.4
twilio==9.2.3
googletrans==4.0.0rc1
